moving addon handlers to addon package to avoid import issues
